updates and new mail function in lib
