+++ functions/package.json
@@ -20,14 +20,8 @@
     "@types/nodemailer": "^6.4.0",
     "@types/nodemailer-smtp-transport": "^2.7.4",
     "firebase-admin": "^8.10.0",
-<<<<<<< HEAD
-    "firebase-functions": "^3.6.0",
-    "luxon": "^1.22.2",
-    "nodemailer": "^6.4.6"
-=======
     "firebase-functions": "^3.6.1",
     "luxon": "^1.22.2"
->>>>>>> 6832b4ab916bdc4875db9efa90ccaab22159b719
   },
   "devDependencies": {
     "@firebase/firestore-types": "^1.10.1",
